refactor-002: 컴포넌트로 분리
