+++ src/components/layout/sideBar/UserMode.tsx
@@ -1,0 +1,5 @@
+const UserMode = () => {
+  return <div></div>;
+};
+
+export default UserMode;
