+++ src/stores/user.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const userState = atom({
+  key: 'user',
+  default: false,
+});
